[feat] introduct Express router to products.js
