+++ mission_3/main.js
@@ -1,14 +1,12 @@
 import express from 'express';
-import { PrismaClient } from '@prisma/client';
-
-const app = express();
-const prisma = new PrismaClient();
+import productRouter from './routes/products.js';
 
 const PORT = process.env.PORT || 3000;
 
+const app = express();
 app.use(express.json());
 
-function logRequest(req, res, next) {
+function logRequest(req, _, next) {
   console.log(`[${req.method}] ${req.path}`)
   next();
 }
@@ -16,158 +14,6 @@ function logRequest(req, res, next) {
 // middleware for loggin all http request
 app.use(logRequest);
 
-// Upload a new product
-app.post('/product', async (req, res) => {
-
-  // destructuring field data from request body
-  const {
-    name,
-    description,
-    price,
-    tags
-  } = req.body;
-
-  // validation logic (possible improvements with Zod library in the future)
-  if ( !name || !description || !price || !tags )
-    return res.status(400).json({ message: "Invalid SQL Parameters" });
-
-  try {
-    // insert into Product table
-    const product = await prisma.product.create({
-      data: {
-        name,
-        description,
-        price,
-        tags
-      }
-    });
-
-    // send 201 response status code indicating the query was processed successfully
-    // with the Product record
-    res.status(201).json(product);
-  } catch (err) {
-    console.error('An error has occurred: ', err);
-
-    res.status(500).json({ message: "An error has occurred during processing sql" });
-  }
-
-});
-
-// Get informations of a certain product
-app.get('/product/:id', async (req, res) => {
-  const { id } = req.params;
-  if (!id)
-    return res.status(400).json({ message: "Invalid parameter 'id'" });
-
-  try {
-    const product = await prisma.product.findUnique({
-      where: {
-        id: parseInt(id)
-      }
-    });
-
-    if (product)
-      res.status(200).json(product);
-    else 
-      res.status(404).json({ message: `Cannot find product with ID ${id}` })
-  } catch (err) {
-    console.error('An error has occurred: ', err.message);
-
-    res.status(500).json({ message: "An error has occurred during processing sql" });
-  }
-});
-
-// Modify a product property
-app.patch('/product/:id', async (req, res) => {
-  const { id } = req.params;
-  if (!id)
-    return res.status(400).json({ message: "Invalid parameter 'id'" });
-
-  // Columns in model Product
-  const productCols = [
-    "name",
-    "description",
-    "price",
-    "tags"
-  ];
-
-  const filteredBody = Object.entries(req.body)
-    .filter(e => productCols.includes(e[0]))
-    .reduce((obj, ele) => {
-      obj[ele[0]] = ele[1]
-      return obj;
-    }, {});
-  
-  try {
-    const product = await prisma.product.update({
-      where: {
-        id: Number(id)
-      },
-      data: {
-        ...filteredBody
-      }
-    });
-
-    if (product)
-      res.status(200).json(product);
-    else 
-      res.status(404).json({ message: `Cannot find product with ID ${id}` });
-  } catch (err) {
-    console.error('An error has occurred: ', err.message);
-
-    res.status(500).json({ message: "An error has occurred during processing sql" });
-  }
-});
-
-// Look up a particular product
-app.delete('/product/:id', async (req, res) => {
-  const { id } = req.params;
-  if (!id)
-    return res.status(400).json({ message: "Invalid parameter 'id'" });
-
-  try {
-    const deleted = await prisma.product.delete({
-      where: {
-        id: Number(id)
-      }
-    });
-
-    if (deleted)
-      res.status(200).json(deleted);
-    else 
-      res.status(404).json({ message: `Cannot find product with ID ${id}` });
-  } catch (err) {
-    console.error('An error has occurred: ', err.message);
-
-    res.status(500).json({ message: "An error has occurred during processing sql" });
-  }
-});
-
-// Inquiry all products
-app.get('/product', async (req, res) => {
-  const { offset = 0, limit = 10 } = req.query;
-
-  try {
-    const products = await prisma.product.findMany({
-      select: {
-        id: true,
-        name: true,
-        price: true,
-        createdAt: true
-      },
-      skip: offset * limit,
-      take: limit
-    });
-
-    if (products)
-      res.status(200).json(products);
-    else 
-      res.status(404).json({ message: `Cannot find product with ID ${id}` });
-  } catch (err) {
-    console.error('An error has occurred: ', err.message);
-
-    res.status(500).json({ message: "An error has occurred during processing sql" });
-  }
-});
+app.use('/products', productRouter);
 
 app.listen(PORT, () => console.log(`Server started on port ${PORT}..`));
